Refactor stats component folder and added tabs to the stats page
