+++ client/src/components/Stats/FriendComp.js
@@ -2,7 +2,7 @@ import React from 'react'
 import { withStyles } from '@material-ui/core/styles'
 import Grid from '@material-ui/core/Grid'
 // import Paper from '@material-ui/core/Paper'
-import FriendOneStatsTable from './FriendOneStatsTable';
+import FriendStatsTable from './FriendStatsTable';
 
 const styles = theme => ({
   root: {
@@ -14,16 +14,16 @@ const styles = theme => ({
   }
 })
 
-function FriendOneComp(props) {
+function FriendComp(props) {
   const { user } = props
   // Deleted classes out props because it was unused
   return (
     <Grid item xs={6} sm container>
-      <FriendOneStatsTable
+      <FriendStatsTable
         user={user}
       />
     </Grid>
   )
 }
 
-export default withStyles(styles)(FriendOneComp)
+export default withStyles(styles)(FriendComp)
